Change MeshAnimation duration and ticksPerSecond types from float to double

+++ libBuild/hdr/buffers/meshanimation.hpp
@@ -9,8 +9,8 @@ namespace gfx
     struct MeshAnimation
     {
         std::string name;
-        float duration;
-        float ticksPerSecond;
+        double duration;
+        double ticksPerSecond;
         std::vector<std::pair<std::string, std::vector<glm::mat4>>> channels;
     };
 }
